add pseudo code for cong-clos
