+++ src/Language/Fixpoint/Congruence/Types.hs
@@ -3,7 +3,20 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Fixpoint.Graph.Congruence where
+module Language.Fixpoint.Congruence.Types 
+  ( -- * Queries 
+    CongQuery (..)
+  , Equality (..)
+  , Disequality (..)
+    
+    -- * Terms 
+  , Term 
+  , identity 
+
+    -- * Constructors
+  , app
+  , var
+  ) where
 
 import Data.Function (on)
 import Data.Hashable
@@ -14,68 +27,14 @@ import qualified Language.Fixpoint.Types as F
 -- 1.  x = y => f x = f y
 -- 2.  f(f(f(x))) = x => f(f(f(f(f(x))))) = x => f(x) = a 
 
-examples = [t1, t2]
+_examples = [t1, t2]
   where 
     t1   = app "f" [var "x", var "y"] 
     t2   = app "f" [var "x", var "y"] 
 
-{-
-
-data UFSt = U 
-  { ufPar   :: M.HashMap Id Term          -- ^ i :-> t      IF term 'i' has union-find parent 't'
-  , ufPreds :: M.HashMap Id [(Int, Term)] -- ^ i :-> (j, t) IF term 'i' is 'j'-th arg in application 't'
-  }
-
-parent :: Term -> UF (Maybe Term)
-parent u = M.lookup (identity u) <$> gets ufPar
-
-find :: Term -> UF Id 
-find u = do 
-  mbU' <- parent u 
-  case mbU' of
-    Nothing -> return u 
-    Just u' -> find u'  
-
-union :: Term -> Term -> UF () 
-union u v = TODO 
-
-preds :: Id -> UF [Term]
-preds = TODO 
-
-
-isCong :: Term -> Term -> UF Bool 
-isCong = TODO 
-
-isEq :: Term -> Term -> UF Bool 
-isEq u v = do 
-  ui <- find u 
-  vi <- find v 
-  return (ui == vi)
-
-congMerge :: (Int, Term) -> (Int, Term) -> UF ()
-congMerge (i, u) (j, v) 
-  | i == j = ifM (isEq u v) 
-              (return ())
-              (ifM (isCong u v) 
-                (merge u v)
-                (return ())
-  | otherwise = return ()  
-
-merge :: Term -> Term -> UF ()
-merge u v = do 
-  ui <- find u
-  vi <- find v 
-  if ui == vi 
-    then return ()
-    else do 
-      uPs <- preds ui 
-      vPs <- preds vi 
-      union u v
-      forM_ uPs $ \u' ->
-        forM_ vPs $ \v' -> 
-          congMerge u' v'
- -}    
-
+data CongQuery   = Query [Equality] [Disequality]  
+data Equality    = Eq    Term Term
+data Disequality = Diseq Term Term
 
 --------------------------------------------------------------------------------
 -- | Exported Constructors
